aws-nemotron-mfg-cowork/requirements.txt
streamlit>=1.32.0
boto3>=1.34.0
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.24.0
